Handle undefined default environment in python client (#2016)
